[ADT] Move llvm::is_detected to STLForwardCompat.h
llvm::is_detected forward-ported from C++20.  As such, it belongs to
STLForwardCompat.h.
